% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split2df_utils.R
\name{header_cleaning}
\alias{header_cleaning}
\title{str_split2df}
\usage{
header_cleaning(header_vector, pattern = "/")
}
\arguments{
\item{header_vector}{string vector}

\item{pattern}{split pattern, Default: '/'}
}
\value{
a splited dataframe
}
\description{
split the strings into dataframe without autofill
}
\details{
DETAILS
}
\examples{
header_cleaning(c('a_b_c'), pattern='_')
}
