% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_ident.R
\name{ORF_ident}
\alias{ORF_ident}
\title{ORF_ident}
\usage{
ORF_ident(
  fasta_path,
  coor_gap_thres = 86,
  out_fasta_path = paste(getwd(), "/", sep = ""),
  out_fasta_prefix = "ORF"
)
}
\arguments{
\item{fasta_path}{direction contain the aligned fasta file}

\item{coor_gap_thres}{coordinate gap proportion up limitation, Default: 86}

\item{out_fasta_path}{identified ORF write direction, Default: working direction}

\item{out_fasta_prefix}{identified ORF file prefix, Default: ORF}
}
\value{
write the open reading frame out
}
\description{
identiifed Open reading frame
}
\details{
DETAILS
}
