% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organ_subtype_utils.R
\name{organ_subtype}
\alias{organ_subtype}
\title{organ_subtype}
\usage{
organ_subtype(in_seq)
}
\arguments{
\item{in_seq}{dataframe contain the subtype and segment information}

\item{subtype_col}{The col contain the segment information}
}
\value{
list contain the internal gene, HA and NA dataframe
}
\description{
Devide the dataframe by subtypes
}
\details{
DETAILS
}
