% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhloprocessing-package.R
\docType{package}
\name{pyhloprocessing-package}
\alias{pyhloprocessing}
\alias{pyhloprocessing-package}
\title{pyhloprocessing: Sequences processing and lineage cleavage in phlyogenetic tree}
\description{
Provides tools for analyzing and visualizing data related to avian influenza. The package includes functions for sequence alignment, phylogenetic analysis, and clade identification, integrating data from GISAID and IRD.
}
\author{
\strong{Maintainer}: Eric Chen \email{cysh010479@gmail.com}

}
\keyword{internal}
