% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_processing.R
\name{raw_processing}
\alias{raw_processing}
\title{raw_processing}
\usage{
raw_processing(
  meta_path,
  gisaid_seq_path,
  ird_seq_path,
  length_thres = 0.8,
  Amb_prop = 5,
  out_fasta_path,
  out_fasta_prefix
)
}
\arguments{
\item{meta_path}{GISAID metadata direction}

\item{gisaid_seq_path}{GISAID sequence direction}

\item{ird_seq_path}{IRD metadata direction}

\item{length_thres}{sequence length threshold, x-fold of median genome length, Default: 0.8}

\item{Amb_prop}{sequence ambiguous letter proportion, Default: 5}

\item{out_fasta_path}{Output direction path}

\item{out_fasta_prefix}{Output file prefix}
}
\value{
write fasta and log file out
}
\description{
merge and filter the genome length, ambiguous character
}
\details{
DETAILS
}
