% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_as_list_utils.R
\name{read_as_list}
\alias{read_as_list}
\title{read_as_list}
\usage{
read_as_list(
  path,
  prefix = "",
  sep = "auto",
  header = "auto",
  file_type = "txt"
)
}
\arguments{
\item{path}{The direction of the files}

\item{prefix}{The prefix of the file which would like to import together, Default: ''}

\item{sep}{The csv/txt separation symbol, Default: 'auto'}

\item{header}{Whether the colnames or not, Default: 'auto'}

\item{file_type}{Which type of files would like to import together, Default: 'txt'}
}
\value{
A list containing the imported files
}
\description{
Read multiple files
}
\details{
DETAILS
}
