% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mafft.R
\name{run_mafft}
\alias{run_mafft}
\title{run_mafft}
\usage{
run_mafft(input_path)
}
\arguments{
\item{input_path}{the direction contain the fasta files for alignment, the file have to be .fa or .fasta}
}
\value{
aligned fasta file in the offered direction
}
\description{
perform alignment through MAFFT
}
\details{
DETAILS
}
