% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_DF_utils.R
\name{table_DF}
\alias{table_DF}
\title{table_DF}
\usage{
table_DF(x, percent = F)
}
\arguments{
\item{x}{data vector}

\item{percent}{whether to percentage, Default: F}
}
\value{
frequency dataframe
}
\description{
Build the frequency table into dataframe
}
\examples{
table_DF(c(1, 1, 1, 2, 2))
}
