% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta_utils.R
\name{write_fasta}
\alias{write_fasta}
\title{write_fasta}
\usage{
write_fasta(sequence, header, file_out)
}
\arguments{
\item{sequence}{sequences vector}

\item{header}{header vector}

\item{file_out}{generated fasta file direction}
}
\description{
Generate Fasta file local file
}
