% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhloprocessing-package.R
\docType{package}
\name{pyhloprocessing-package}
\alias{pyhloprocessing}
\alias{pyhloprocessing-package}
\title{pyhloprocessing: Sequences processing and lineage cleavage in phlyogenetic tree}
\description{
Package process sequences for phylogenetics and lineage cleavage
}
\author{
\strong{Maintainer}: Eric Chen \email{cysh010479@gmail.com}

}
\keyword{internal}
